refactor: update imports to use consolidated search context
