+++ frontend/src/components/TabContainer.tsx
@@ -1,7 +1,7 @@
 import { useRef, useEffect, useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { useSearchState } from "@/search/context/SearchStateContext";
+import { useSearchState } from "@/search/context";
 
 import { Song } from "../types/song";
 import type { StoredChordSheet } from "@/storage/types";
