+++ frontend/src/search/context/SearchStateContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { SearchStateContextValue } from "./SearchStateContext.types";
+
+export const SearchStateContext = createContext<
+  SearchStateContextValue | undefined
+>(undefined);
