+++ frontend/src/search/context/defaultSearchState.ts
@@ -1,0 +1,7 @@
+import type { SearchState } from "./SearchStateContext.types";
+
+export const defaultSearchState: SearchState = {
+  artist: "",
+  song: "",
+  results: [],
+};
