+++ frontend/src/search/hooks/use-search-redirect.ts
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
-import { useSearchState } from "@/search/context/SearchStateContext";
+import { useSearchState } from "@/search/context";
 import { fromSlug } from "@/utils/url-slug-utils";
 
 // Restores search state from URL parameters on mount
@@ -12,14 +12,14 @@ export function useSearchRedirect() {
     // Only handle search routes
     if (location.pathname === "/search") {
       const searchParams = new URLSearchParams(location.search);
-      const artistParam = searchParams.get('artist');
-      const songParam = searchParams.get('song');
+      const artistParam = searchParams.get("artist");
+      const songParam = searchParams.get("song");
 
       // If URL has search parameters, restore the search state
       if (artistParam || songParam) {
-        const artist = artistParam ? fromSlug(artistParam) : '';
-        const song = songParam ? fromSlug(songParam) : '';
-        
+        const artist = artistParam ? fromSlug(artistParam) : "";
+        const song = songParam ? fromSlug(songParam) : "";
+
         updateSearchState({ artist, song, results: [] });
       }
     }
